Move last methods from LayoutController to LayoutHelper, remove empty LayoutController class
